moved CellEditOperation from common to browser
